chore(reflector): Split the reflector by command
The Reflector was a bit crowded with responsibilities which made it
quite hard to follow esp after coming back to the code after a while.
