+++ lib/fakerbot/command.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'forwardable'
-require_relative 'reflector'
+
 require_relative 'renderer'
 
 module FakerBot
@@ -12,7 +12,7 @@ module FakerBot
     attr_reader :options
 
     def render(result, output)
-      FakerBot::Renderer.call(result, options, output)
+      Renderer.call(result, options, output)
     end
   end
 end
