+++ lib/fakerbot/reflector.rb
@@ -3,8 +3,11 @@
 require 'faker'
 
 module FakerBot
-  # Exposes `Faker` reflection methods
+  # Abstract `Faker` Reflector - introspects the `Faker` module
+  #
   # @api private
+  # @abstract
+  #
   class Reflector
     Faker::Base.class_eval do
       # Select `Faker` subclasses
@@ -22,56 +25,17 @@ module FakerBot
       end
     end
 
-    attr_reader :descendants_with_methods, :query
+    attr_reader :descendants_with_methods
 
-    def initialize(query = nil)
+    def initialize(*)
       @descendants_with_methods = Hash.new { |h, k| h[k] = [] }
-      @query = query
     end
 
-    class << self
-      def find(query)
-        new(query).find
-      end
-
-      def list(show_methods: false)
-        new.list(show_methods)
-      end
-    end
-
-    def find
-      search_descendants_matching_query
-      descendants_with_methods
-    end
-
-    def list(show_methods)
-      show_methods ? all_descendants_with_methods : faker_descendants
-    end
-
-    private
-
-    def all_descendants_with_methods
-      faker_descendants.each do |faker|
-        store(faker, faker.my_singleton_methods)
-      end
-      descendants_with_methods
-    end
-
-    def search_descendants_matching_query
-      faker_descendants.each do |faker|
-        methods = faker.my_singleton_methods
-        matching = methods.select { |m| query_matches?(m.to_s) }
-        store(faker, matching)
-      end
-    end
-
-    def query_matches?(method_name)
-      method_name_parts = method_name.split(/_/).reject(&:empty?)
-      query.match(/#{method_name_parts.join('|')}/)
-    end
+    protected
 
     def store(descendant, methods)
       return if methods.empty?
+
       descendants_with_methods[descendant].concat(methods)
     end
