+++ lib/fakerbot/reflectors/search.rb
@@ -1,0 +1,44 @@
+# frozen_string_literal: true
+
+require_relative '../reflector'
+
+module FakerBot
+  module Reflectors
+    # Find command reflector
+    # @api private
+    #
+    class Search < FakerBot::Reflector
+      attr_reader :query
+
+      def self.call(query)
+        new(query).call
+      end
+
+      def initialize(query)
+        @query = query
+
+        super
+      end
+
+      def call
+        search_descendants_matching_query
+        descendants_with_methods
+      end
+
+      private
+
+      def search_descendants_matching_query
+        faker_descendants.each do |descendant|
+          methods = descendant.my_singleton_methods
+          matching = methods.select { |m| query_matches?(m.to_s) }
+          store(descendant, matching)
+        end
+      end
+
+      def query_matches?(method_name)
+        method_name_parts = method_name.split(/_/).reject(&:empty?)
+        query.match(/#{method_name_parts.join('|')}/)
+      end
+    end
+  end
+end
